# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "log-viewer API",
    "description": "REST API for listing and reading log files under LOG_ROOT.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host (Flask dev server)"
    }
  ],
  "paths": {
    "/api/logs": {
      "get": {
        "summary": "List log files",
        "description": "Returns sorted log entries in the configured log directory. Each item includes `name`, `issues` (`error` / `warning` when the corresponding counts are positive), `error_count` and `warning_count` from the same sampled scan (approximate for large files).",
        "operationId": "listLogs",
        "responses": {
          "200": {
            "description": "Directory path and file names",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["root", "files"],
                  "properties": {
                    "root": { "type": "string", "description": "Absolute path to log directory" },
                    "files": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["name", "issues", "error_count", "warning_count"],
                        "properties": {
                          "name": { "type": "string", "description": "File name under the log root" },
                          "issues": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "enum": ["error", "warning"]
                            },
                            "description": "Present when the sample had matches for that severity"
                          },
                          "error_count": {
                            "type": "integer",
                            "minimum": 0,
                            "description": "Occurrences of error-level heuristic in sample"
                          },
                          "warning_count": {
                            "type": "integer",
                            "minimum": 0,
                            "description": "Occurrences of warning heuristic in sample"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Log root missing or not a directory",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/logs/hourly-summary": {
      "get": {
        "summary": "Per-file hourly health for a day and trailing 24-day window",
        "description": "Single pass per file: (1) hours 0–23 for `date`; (2) 24 calendar days in `days24` with index 23 = `date`, 22 = previous day, …, 0 = oldest. Each slot is `empty`, `ok`, `warning`, `error`, or `error_warning` (both heuristics matched in that slot; UI may show a diagonal split).",
        "operationId": "hourlySummary",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" },
            "description": "Calendar day in the viewer's local timezone (typically “today”)"
          }
        ],
        "responses": {
          "200": {
            "description": "Date key and per-file hour arrays",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["date", "recent_days", "files"],
                  "properties": {
                    "date": { "type": "string", "description": "Echo of the requested YYYY-MM-DD (newest day = `days24` index 23)" },
                    "recent_days": { "type": "integer", "const": 24, "description": "Length of each `days24` array" },
                    "files": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["name", "hours", "days24"],
                        "properties": {
                          "name": { "type": "string" },
                          "hours": {
                            "type": "array",
                            "minItems": 24,
                            "maxItems": 24,
                            "items": {
                              "type": "string",
                              "enum": ["empty", "ok", "warning", "error", "error_warning"]
                            }
                          },
                          "days24": {
                            "type": "array",
                            "minItems": 24,
                            "maxItems": 24,
                            "items": {
                              "type": "string",
                              "enum": ["empty", "ok", "warning", "error", "error_warning"]
                            },
                            "description": "Index 23 = `date`; 22 = previous calendar day; 0 = 23 days before `date`"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or missing `date`",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Log root missing or not a directory",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/logs/{name}": {
      "get": {
        "summary": "Get log file content",
        "description": "Reads a single file by name (no path segments). Large files may be truncated; use `tail` to limit lines.",
        "operationId": "getLog",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "File name only (e.g. docker.log)"
          },
          {
            "name": "tail",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1 },
            "description": "If set, only the last N lines are returned (after any byte truncation)."
          }
        ],
        "responses": {
          "200": {
            "description": "File metadata and text content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["name", "path", "size", "truncated", "content"],
                  "properties": {
                    "name": { "type": "string" },
                    "path": { "type": "string" },
                    "size": { "type": "integer" },
                    "truncated": { "type": "boolean" },
                    "content": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "File not found or invalid name",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Filesystem error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/logs/{name}/analyze-presets": {
      "get": {
        "summary": "AI analysis preset sizes for one log",
        "description": "Returns approximate token counts (UTF-8 bytes of time-filtered slice / 3) per preset. Large files use the same tail cap as GET /api/logs/{name}; stats apply to that slice only.",
        "operationId": "analyzePresets",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Per-preset byte and token estimates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["name", "source_size", "source_truncated", "presets"],
                  "properties": {
                    "name": { "type": "string" },
                    "source_size": { "type": "integer" },
                    "source_truncated": { "type": "boolean" },
                    "presets": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["id", "label", "bytes", "approx_tokens"],
                        "properties": {
                          "id": {
                            "type": "string",
                            "enum": ["last_1h", "today", "week", "last_7d", "last_30d"]
                          },
                          "label": { "type": "string" },
                          "bytes": { "type": "integer", "minimum": 0 },
                          "approx_tokens": { "type": "integer", "minimum": 0 }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Filesystem error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/logs/{name}/analyze": {
      "post": {
        "summary": "Run AI analysis for a time window",
        "description": "Proxies log excerpt to the configured AI API (default POST http://127.0.0.1:8500/analyze-log). Set AI_ANALYZE_URL to override.",
        "operationId": "analyzeLog",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["preset"],
                "properties": {
                  "preset": {
                    "type": "string",
                    "enum": ["last_1h", "today", "week", "last_7d", "last_30d"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Analysis text from AI API",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["analysis", "preset", "approx_tokens", "profile", "ai_error", "source_truncated"],
                  "properties": {
                    "analysis": { "type": "string" },
                    "preset": { "type": "string" },
                    "approx_tokens": { "type": "integer" },
                    "profile": { "type": "string" },
                    "ai_error": { "type": ["string", "null"] },
                    "source_truncated": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid preset",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "502": {
            "description": "AI API unreachable or error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "analysis": { "type": "string" },
                    "ai_error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  }
}